Move depletion tests into normal regression/unit test directories
