+++ tests/regression_tests/test_deplete_full.py
